+++ packages/zp-plugins/src/hooks/all/queryHook.ts
@@ -4,7 +4,7 @@ import { HookContext } from '..'
 const queryHook = <T, C extends HookContext>(resolver: Resolver<T, C>) => {
   return async (ctx: C) => {
     if (ctx.params?.query) {
-      ctx.params.query = resolver.resolve(ctx.params.query as any, ctx)
+      ctx.params.query = await resolver.resolve(ctx.params.query as any, ctx) as any
     }
   }
 }