+++ packages/zp/CHANGELOG.md
@@ -1,5 +1,11 @@
 # zapnode
 
+## 0.0.15
+
+### Patch Changes
+
+- update queryHook and change hooks to execute in sequence
+
 ## 0.0.14
 
 ### Patch Changes